inference-server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
transformers>=4.47.0
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.0.0
accelerate>=0.26.0
python-multipart==0.0.20
